requirements.txt
opencv-python
python-dotenv
langchain-community
langchain
streamlit
langchain-aws